deleteTest virekr

+++ src/main/java/entities/Ingredient.java
@@ -13,7 +13,7 @@ public class Ingredient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private int id;
 
     @Column(name = "name")
     private String name;
@@ -29,8 +29,9 @@ public class Ingredient {
         this.name = name;
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
 //    public List<RecipeIngredient> getRecipeIngredients() {
@@ -52,6 +53,7 @@ public class Ingredient {
                 ", name='" + name + '\'' +
                 '}';
     }
+
 
     @Override
     public boolean equals(Object o) {
